Albufera Rice Paddy Clusters

Datasets containing information on rice field clusters into which the rice production area of the Albufera Natural Park is decomposed. For the actual definition of these clusters, see Martínez-Megías et al. (2023).

Usage

albufera_clusters

albufera_cluster_geometries

Format

albufera_clusters

A data frame with one row per cluster, and the following columns:

cluster_id

character. Unique identifier of cluster.

ditch

character. Ditch to which the cluster pertains.

area

numeric. Area (in squared meters) of the cluster.

tancat

logical. Whether the cluster is a "tancat".

rice_variety

character. Variety of rice planted in the cluster.

case

factor. ...?

albufera_cluster_geometries

A data frame with one row per cluster, and the columns:

cluster_id

character. Unique identifier of cluster.

geometry

vector of class sfc_MULTIPOLYGON.

Source

Martínez-Megías, C., Mentzel, S., Fuentes-Edfuf, Y., Moe, S. J., & Rico, A. (2023). Influence of climate change and pesticide use practices on the ecological risks of pesticides in a protected Mediterranean wetland: A Bayesian network approach. Science of The Total Environment, 878, 163018. https://doi.org/10.1016/j.scitotenv.2023.163018

Details

Both datasets have the same cardinality (one row per cluster), and they are separated merely for efficiency reasons. They can be joined using the common cluster_id key column.