Albufera Global Hydrological Balance

Wrapper around hb_global, used to run the global hydrological balance calculations with the data for the Albufera lake packed up in the data.frames exported by erahumed.

Usage

albufera_hb_global(
  outflows_df = erahumed::albufera_outflows,
  petp_df = erahumed::albufera_petp,
  storage_curve = linear_storage_curve(intercept = 16.7459 * 1e+06, slope = 23.6577 *
    1e+06),
  petp_surface = linear_petp_surface(surface_P = 114.225826072 * 1e+06, surface_ETP =
    79.360993685 * 1e+06)
)

Arguments

outflows_df, petp_df

data.frames, whose structures follow the templates of albufera_outflows and albufera_petp, respectively; see details.

storage_curve

a function that takes a numeric vector as input, and returns a numeric vector of the same length. Function that converts lake levels (passed through the level argument) into lake volumes.

petp_surface

a function that takes two numeric vectors of common length as inputs, and returns a numeric vector of the same length. Function that converts precipitation and evapotranspiration values (passed through the rain_mm and evapotranspiration_mm arguments) into an overall volume change.

Value

Same as hb_global.

Details

The numeric inputs for the linear storage curve are taken from the CHJ report Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.. The values used as the arguments of petp_surface() were calculated by the package authors, and correspond to the total study area (surface_P) and the flooded surface (surface_ETP).

The outflows_df data.frame is supposed to have all columns of albufera_outflows whose names do not start by outflow_, with the appropriate type. In addition, outflows_df can have an arbitrary number of numeric columns named outflow_*, which represent the measured outflows for the system. It is fundamental that outflow columns follow this particualr naming scheme, as these are automatically recognized by albufera_hb_global() and passed down to the outflow argument of hb_global.

Author

Pablo Amador Crespo, Valerio Gherardi