Albufera Local Hydrological Balance

Wrapper around hb_local, used to run the local hydrological balance simulation algorithm with the data for the Albufera lake packed up in the data.frames exported by erahumed.

Usage

albufera_hb_local(
  outflows_df = erahumed::albufera_outflows,
  petp_df = erahumed::albufera_petp,
  management_df = erahumed::albufera_management,
  clusters_df = erahumed::albufera_clusters,
  storage_curve = linear_storage_curve(intercept = 16.7459 * 1e+06, slope = 23.6577 *
    1e+06),
  petp_surface = linear_petp_surface(surface_P = 114.225826072 * 1e+06, surface_ETP =
    79.360993685 * 1e+06),
  date_min = NULL,
  date_max = NULL,
  ideal_flow_rate_cm = 5
)

Arguments

outflows_df, petp_df, management_df, clusters_df

data.frames, whose structures follow the templates of albufera_outflows, albufera_petp, albufera_management and albufera_clusters, respectively.

storage_curve

a function that takes a numeric vector as input, and returns a numeric vector of the same length. Function that converts lake levels (passed through the level argument) into lake volumes.

petp_surface

a function that takes two numeric vectors of common length as inputs, and returns a numeric vector of the same length. Function that converts precipitation and evapotranspiration values (passed through the rain_mm and evapotranspiration_mm arguments) into an overall volume change.

date_min, date_max

date range for the simulation.

ideal_flow_rate_cm

a positive number. Ideal inflow for days such that ideal_irrigation and ideal_draining are both TRUE.

Value

Same as hb_local.

Details

The numeric inputs used from the linear storage curve and P-ETP surface were extracted from the CHJ report Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.