Albufera Management Data

Data on management schedules of rice paddies in the Albufera National Park. Contains information on irrigation, draining and water level of paddies by day of year.

Usage

albufera_management

Format

albufera_management

The dataset has one row per day of year for all combinations of the categorical variables tancat and variety. The columns are the following:

mm

numeric. Month of year (1 = January, 2 = February, etc.).

dd

numeric. Day of month.

tancat

logical. Whether the paddy is a "tancat" or not.

variety

character. Variety of rice planted in the paddy under consideration.

sowing

logical. Whether mm and dd correspond to the sowing day.

irrigation

logical. Whether the paddy is scheduled to be irrigated on this day.

draining

logical. Whether the paddy is scheduled to be drained on this day.

height_cm

numeric. Water level of the paddy.

Source

https://doi.org/10.1016/j.scitotenv.2023.163018