CT: Chemical Transport

This simulation layer computes the evolution of chemicals applied to rice paddy clusters, based on the previously computed simulations for hydrological balance and chemicals application. The result is a set of time series of chemical masses, one for each applied chemical and for the three compartments: foliage, water and sediment.

Usage

setup_ct(
  simulation,
  drift = 0,
  covmax = 0.5,
  jgrow = 152,
  SNK = 0,
  dact_m = 0.1,
  css_ppm = 50,
  foc = 0.17,
  bd_g_cm3 = 1.5,
  qseep_m_day = 0,
  wilting = 0.24,
  fc = 0.35
)

Arguments

simulation

An object of class erahumed_simulation.

drift

A number between 0 and 1. Percentage of chemical applications lost to drift.

covmax

A number between 0 and 1. Interception potential of foliage at crop maturation.

jgrow

A positive integer. Length (in days) of crop maturation cycle.

SNK

A number between 0 and 1. (TODO).

dact_m

A positive number. Active sediment layer depth, expressed in meters.

css_ppm

A positive number. Suspended sediment concentration, expressed in parts per million.

foc

A number between 0 and 1. Fraction of organic content.

bd_g_cm3

A positive number. Bulk density of the sediment, expressed in grams per cubic centimeter.

qseep_m_day

A positive number. Seepage rate, expressed in meters per day.

wilting

A number between 0 and 1. Wilting point.

fc

A number between 0 and 1. Field capacity.

Value

An objects of class erahumed_simulation.

See also