ERAHUMED modeling interface

The programming interface of the {erahumed} package reflects the sequential structure of the ERAHUMED modeling chain. Each step of the sequence is referred across this documentation as a model "component" (or, sometimes, informally as a modeling "layer"). This documentation page focuses on describing the set of abstractions provided by {erahumed} to deal with sequential modeling and model components.

The first abstraction is provided by erahumed_model objects. These objects, technically implemented as S3 classes, are simple containers for model components - that is, a erahumed_model is a list that stores a sequence of model components. The way components of a model are populated is described subsequently.

We define the following model components, whose technical implementation is discussed extensively in the linked documentation pages:

inp: INPut data.

hba: Hydrological Balance of the Albufera lake.

hbp: Hydrological Balance of rice Paddy clusters.

ca: Chemical Applications. The order of model components in the list above is the logical one. Each component depends on the previous ones (referred to as "upstream"), and is a dependency of the subsequent ones (referred to as "downstream").

Each of these components has associated:

A compute_*() function, that takes a erahumed_model as input and returns a model with the desired component computed on top as output.

An extractor function of the form *() (the asterisk standing for the actual component name), that extracts the desired component from a erahumed_model.

An S3 class following the naming scheme erahumed_*, that inherits from erahumed_model_component, with ad-hoc print and plot methods. Components can be re-computed (that is, the input model to compute_*()) can have a previous computation of the same components. In this case, in order to avoid confusing results, any existing downstream dependency of the component to be recomputed is erased from the output model, and must be recomputed if required.

The example code below illustrates typical operations with model components,

Author

Valerio Gherardi

Examples

### Computation workflow
m <- erahumed_model()  # Initialize a blank model

m <- m |> compute_inp() |> compute_hba()  # Compute INP and HBA components

m
#> An ERAHUMED model.
#> Calculated components:  inp, hba
inp(m)
#> An ERAHUMED INP model component.
#> 
#> Output columns: date, level, outflow_pujol, outflow_perellonet, outflow_perello, is_imputed_level, is_imputed_outflow, rain_mm, evapotranspiration_mm
hba(m)
#> An ERAHUMED HBA model component.
#> 
#> Output columns: level, rain_mm, evapotranspiration_mm, date, is_imputed_level, is_imputed_outflow, volume, volume_change, volume_change_petp, outflow_pujol, outflow_perellonet, outflow_perello, outflow_extra, outflow_total, inflow_total, residence_time_days
hbp(m)  # NULL because not yet computed
#> NULL

m <- m |> compute_inp()  # Recompute INP layer
hba(m)  # NULL because of previous line
#> NULL

# The code below results in an error, because the HBP component depends
# on the HBA component, which is still NULL (see above).
if (FALSE) { # \dontrun{
compute_hbp(m)
} # }

### Extracting and plotting outputs
m <- m |> compute_inp() |> compute_hba()

component_output(inp(m)) |> head()  # 'erahumed_model_component' method
#>         date     level outflow_pujol outflow_perellonet outflow_perello
#> 1 2005-12-20 0.3725000      5.739292           2.919471        2.785127
#> 2 2005-12-21 0.3726458      5.901080           3.140746        2.919733
#> 3 2005-12-22 0.3683833      3.877427           2.500297        2.075301
#> 4 2005-12-23 0.3662375      5.410336           3.180986        2.763805
#> 5 2005-12-24 0.3667417      1.055759           1.578490        1.226585
#> 6 2005-12-25 0.3634917      6.922879           3.786494        3.249450
#>   is_imputed_level is_imputed_outflow rain_mm evapotranspiration_mm
#> 1            FALSE               TRUE     8.2                  0.54
#> 2            FALSE               TRUE     0.2                  0.65
#> 3            FALSE               TRUE     0.2                  0.62
#> 4            FALSE               TRUE     0.2                  1.02
#> 5            FALSE               TRUE     0.0                  0.59
#> 6            FALSE               TRUE     0.2                  0.86
component_output(m, "hba") |> head()  # 'erahumed_model' method
#>       level rain_mm evapotranspiration_mm       date is_imputed_level
#> 1 0.3725000     8.2                  0.54 2005-12-20            FALSE
#> 2 0.3726458     0.2                  0.65 2005-12-21            FALSE
#> 3 0.3683833     0.2                  0.62 2005-12-22            FALSE
#> 4 0.3662375     0.2                  1.02 2005-12-23            FALSE
#> 5 0.3667417     0.0                  0.59 2005-12-24            FALSE
#> 6 0.3634917     0.2                  0.86 2005-12-25            FALSE
#>   is_imputed_outflow   volume volume_change volume_change_petp outflow_pujol
#> 1               TRUE 25558393      3450.081          893796837      5.739292
#> 2               TRUE 25561843   -100840.946          -28739481      5.901080
#> 3               TRUE 25461002    -50765.481          -26358651      3.877427
#> 4               TRUE 25410237     11927.424          -58103048      5.410336
#> 5               TRUE 25422164    -76887.525          -46822986      1.055759
#> 6               TRUE 25345277   -157323.705          -45405289      6.922879
#>   outflow_perellonet outflow_perello outflow_extra outflow_total inflow_total
#> 1           2.919471        2.785127      10333.39  10344.830869       0.0000
#> 2           3.140746        2.919733          0.00     11.961558     343.4273
#> 3           2.500297        2.075301          0.00      8.453025     312.9424
#> 4           3.180986        2.763805          0.00     11.355127     683.9822
#> 5           1.578490        1.226585          0.00      3.860835     544.9036
#> 6           3.786494        3.249450          0.00     13.958823     537.6621
#>   residence_time_days
#> 1          0.04163975
#> 2          0.04258338
#> 3          0.04358226
#> 4          0.04463149
#> 5          0.04573047
#> 6          0.04687804

if (FALSE) { # \dontrun{
# Returns an interactive plotly plot, run manually.
plot(hba(m), variable = "outflow_total")
} # }