Setup parameters of ERAHUMED simulations

These functions are used to set up the input parameters of ERAHUMED simulations.

Usage

setup_exposure(
  simulation,
  ca_schedules_df = erahumed::albufera_ca_schedules,
  drift = 0,
  covmax = 0.5,
  jgrow = 152,
  SNK = 0,
  dact_m = 0.1,
  css_ppm = 50,
  foc = 0.17,
  bd_g_cm3 = 1.5,
  qseep_m_day = 0,
  wilting = 0.24,
  fc = 0.35
)

setup_hydrology(
  simulation,
  outflows_df = erahumed::albufera_outflows,
  weather_df = erahumed::albufera_weather,
  variety_prop = c(J.Sendra = 0.8, Bomba = 0.1, Clearfield = 0.1),
  storage_curve = function(level) 25.58 * 1e+06 + level * 63.086 * 1e+06,
  petp_function = function(p, etp) 53.9 * 1000 * (p - etp),
  management_df = erahumed::albufera_management,
  ideal_flow_rate_cm = 5,
  height_thresh_cm = 0.5,
  ditch_level_m = 1,
  seed = 840
)

setup_risk(simulation)

Arguments

simulation

[erahumed_simulation]
The simulation object being modified.

ca_schedules_df

[data.frame]
A dataset that provides the list of scheduled chemical applications, along the template of albufera_ca_schedules.

drift

[numeric(1)]
A number between 0 and 1. Percentage of chemical applications lost to drift.

covmax

[numeric(1)]
A number between 0 and 1. Interception potential of foliage at crop maturation.

jgrow

[numeric(1)]
A positive integer. Length (in days) of crop maturation cycle.

SNK

[numeric(1)]
A number between 0 and 1. (TODO).

dact_m

[numeric(1)]
A positive number. Active sediment layer depth, expressed in meters.

css_ppm

[numeric(1)]
A positive number. Suspended sediment concentration, expressed in parts per million.

foc

[numeric(1)]
A number between 0 and 1. Fraction of organic content.

bd_g_cm3

[numeric(1)]
A positive number. Bulk density of the sediment, expressed in grams per cubic centimeter.

qseep_m_day

[numeric(1)]
A positive number. Seepage rate, expressed in meters per day.

wilting

[numeric(1)]
A number between 0 and 1. Wilting point.

fc

[numeric(1)]
A number between 0 and 1. Field capacity.

outflows_df

[data.frame]
Time-series dataset that provides the observational hydrological data on the Albufera lake, along the template of albufera_outflows (the default value).

weather_df

[data.frame]
A dataset that provides the relevant metereological time series, along the template of albufera_weather (the default value).

variety_prop

[numeric(3)]
Vector of 3 positive values. Controls the proportions of field surface allocated to the cultivation of the J. Sendra, Bomba, and Clearfield varieties, respectively (positive vectors that do not sum to one are allowed, with entries being understood as relative proportions).

storage_curve

[function]
A function that converts lake levels into lake volumes. The function should take a single numeric vector as input, and should return a numeric vector of the same length as output. The numeric values for the default (linear) storage curve are taken from the CHJ report Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.

petp_function

[function]
A function that converts values of precipitation and evapotranspiration per unit area (assumed to be expressed in millimiters) into volume changes for the Albufera lake (in cube meters when the inputs are given in the expected units). This function should take two numeric vectors of common length as inputs (representing the precipitation and evapotranspiration time series, respectively), and should return a numeric vector of the same length. The numeric values used in the default (linear) function were computed by the package authors, and correspond to the total study area (consisting of the lake and irrigation ditches) and the flooded fraction, that are relevant for precipitation and evapotranspiration volume changes, respectively.

management_df

[data.frame]
Dataset that provides the yearly schedule for irrigation and draining, along the template of albufera_management (the default value).

ideal_flow_rate_cm

[numeric(1)]
Ideal inflow/outflow of a cluster, for days in which the cluster is scheduled to be in flux (i.e. when being simultaneously irrigated and drained). Expressed in centimeters per day.

height_thresh_cm

[numeric(1)]
A positive number. Height threshold for water levels, below which a cluster is considered to be emptied.

ditch_level_m

[numeric(1)]
Constant water depth in ditches.

seed

[numeric(1)]
Seed for random number generation used by the simulation algorithms.

Value

A erahumed_simulation.