ERAHUMED simulation interface

The programming interface of the {erahumed} package reflects the sequential structure of the ERAHUMED simulation chain. Each step of the sequence is referred across this documentation as a simulation "layer". This documentation page focuses on describing the set of abstractions provided by {erahumed} to deal with simulation layers. For further information, you can consult the main package vignette.

The first abstraction is provided by erahumed_simulation objects. These objects, technically implemented as S3 classes, are simple containers for simulation layers - that is, a erahumed_simulation is a list that stores simulation layers. In turn, layers are implemented as S3 classes following the naming scheme erahumed_*, that inherit from erahumed_simulation_layer, with ad-hoc print and plot methods.

Creating a new erahumed_simulation object will setup, but not run, a fresh new simulation with default settings for all the above layers. Each layer has associated a setup_*() function that allows to modify layer-specific configurations. In order to run (i.e compute the actual results) of a simulation, one would use run_simulation.

We define the following layers, whose technical implementation is discussed extensively in the linked documentation pages:

inp: INPut data.

hba: Hydrological Balance of the Albufera lake.

hbp: Hydrological Balance of rice Paddy clusters.

ca: Chemical Applications.

ct: Chemical Transport.

The order of simulation layers in the list above is the logical one. Each layer depends on the previous ones (referred to as "upstream"), and is a dependency of the subsequent ones (referred to as "downstream"). The list of ERAHUMED layers can also be obtained through the erahumed_layers helper.

The example code below illustrates typical operations with simulation layers.

Examples

# Simulation workflow

### Initialize a new simulation
s <- erahumed_simulation()

### Fine-tune layers
s <- s |>
  setup_hbp(ideal_flow_rate_cm = 2.5) |>
  setup_ct(dact_m = 0.1)

### Run simulation until a given layer
s <- run_simulation(s, layer = "hba")

### Extract layers
get_layer(s, "inp")
#> A ERAHUMED INP simulation layer.
#> 
#> Output columns: date, level, outflow_pujol, outflow_perellonet, outflow_perello, is_imputed_level, is_imputed_outflow, temperature_ave, temperature_min, temperature_max, precipitation_mm, evapotranspiration_mm
get_layer_output(s, "hba") |> head()
#>       level precipitation_mm evapotranspiration_mm       date is_imputed_level
#> 1 0.3725000              8.2                  0.54 2005-12-20            FALSE
#> 2 0.3726458              0.2                  0.65 2005-12-21            FALSE
#> 3 0.3683833              0.2                  0.62 2005-12-22            FALSE
#> 4 0.3662375              0.2                  1.02 2005-12-23            FALSE
#> 5 0.3667417              0.0                  0.59 2005-12-24            FALSE
#> 6 0.3634917              0.2                  0.86 2005-12-25            FALSE
#>   is_imputed_outflow   volume volume_change volume_change_petp outflow_pujol
#> 1               TRUE 25558393      3450.081          893798.26      5.739292
#> 2               TRUE 25561843   -100840.946          -28739.45      5.901080
#> 3               TRUE 25461002    -50765.481          -26358.62      3.877427
#> 4               TRUE 25410237     11927.424          -58103.02      5.410336
#> 5               TRUE 25422164    -76887.525          -46822.99      1.055759
#> 6               TRUE 25345277   -157323.705          -45405.26      6.922879
#>   outflow_perellonet outflow_perello outflow_recirculation outflow_total
#> 1           2.919471        2.785127                     0     11.443889
#> 2           3.140746        2.919733                     0     11.961558
#> 3           2.500297        2.075301                     0      8.453025
#> 4           3.180986        2.763805                     0     11.355127
#> 5           1.578490        1.226585                     0      3.860835
#> 6           3.786494        3.249450                     0     13.958823
#>   inflow_total residence_time_days
#> 1     1.138934            21.01039
#> 2    11.127050            20.97162
#> 3     8.170538            20.97415
#> 4    12.165665            20.92734
#> 5     3.512866            20.92254
#> 6    12.663471            20.86514
get_layer_output(s, "hbp")  # NULL because not yet computed
#> NULL
get_layer_parameters(s, "ca")  # Layer parameters set during initialization
#> $ca_schedules_df
#> # A tibble: 29 × 5
#>      day rice_variety chemical    kg_per_ha application_type
#>    <dbl> <chr>        <chr>           <dbl> <chr>           
#>  1    51 J.Sendra     Acetamiprid      0.03 ground          
#>  2    51 Bomba        Acetamiprid      0.03 ground          
#>  3    51 Clearfield   Acetamiprid      0.03 ground          
#>  4    51 J.Sendra     Benta            1    ground          
#>  5    51 Bomba        Benta            1    ground          
#>  6    51 J.Sendra     MCPA             0.5  ground          
#>  7    51 Bomba        MCPA             0.5  ground          
#>  8    21 J.Sendra     Penoxulam        0.04 ground          
#>  9    21 Bomba        Penoxulam        0.04 ground          
#> 10     7 J.Sendra     Cyhalo           0.3  ground          
#> # ℹ 19 more rows
#> 

### Reconfigure a layer
s <- s |> setup_hba(storage_curve = \(level) level + 1)
get_layer_output(s, "hba")  # NULL because invalidated by previous line...
#> NULL
s <- run_simulation(s, layer = "hba")  # ... which requires to re-run layer
get_layer_output(s, "hba") |> head()
#>       level precipitation_mm evapotranspiration_mm       date is_imputed_level
#> 1 0.3725000              8.2                  0.54 2005-12-20            FALSE
#> 2 0.3726458              0.2                  0.65 2005-12-21            FALSE
#> 3 0.3683833              0.2                  0.62 2005-12-22            FALSE
#> 4 0.3662375              0.2                  1.02 2005-12-23            FALSE
#> 5 0.3667417              0.0                  0.59 2005-12-24            FALSE
#> 6 0.3634917              0.2                  0.86 2005-12-25            FALSE
#>   is_imputed_outflow   volume volume_change volume_change_petp outflow_pujol
#> 1               TRUE 1.372500  0.0001458333          893798.26      5.739292
#> 2               TRUE 1.372646 -0.0042625000          -28739.45      5.901080
#> 3               TRUE 1.368383 -0.0021458333          -26358.62      3.877427
#> 4               TRUE 1.366237  0.0005041667          -58103.02      5.410336
#> 5               TRUE 1.366742 -0.0032500000          -46822.99      1.055759
#> 6               TRUE 1.363492 -0.0066500000          -45405.26      6.922879
#>   outflow_perellonet outflow_perello outflow_recirculation outflow_total
#> 1           2.919471        2.785127                     0     11.443889
#> 2           3.140746        2.919733                     0     11.961558
#> 3           2.500297        2.075301                     0      8.453025
#> 4           3.180986        2.763805                     0     11.355127
#> 5           1.578490        1.226585                     0      3.860835
#> 6           3.786494        3.249450                     0     13.958823
#>   inflow_total residence_time_days
#> 1     1.099002        1.116036e-06
#> 2    12.294191        1.114549e-06
#> 3     8.758102        1.115188e-06
#> 4    12.027616        1.113261e-06
#> 5     4.402768        1.113570e-06
#> 6    14.484347        1.111186e-06

if (FALSE) { # \dontrun{
# Returns an interactive plotly plot, run manually.
s |> get_layer("hba") |> plot(variable = "outflow_total")
} # }