HBP: Hydrological Balance of rice Paddies

This simulation layer performs a simulation of hydrological balance daily data for the rice paddies of the Albufera National Park, that provides a synthetic dataset of daily inflows, outflows and average water levels for each rice paddy.

The inputs to this simulation are:

The previously calculated hydrological balance for the Albufera lake.

A geographical subdivision of the Albufera National Park into "clusters" of rice paddies, of a definite rice variety, and whose waters flow into the same ditch.

An yearly ideal schedule for irrigation and draining, with corresponding expected water levels of a cluster for each day of the year.

The data for the second input is stored internally in {erahumed}, and can be examined through the clusters helper.

Usage

setup_hbp(
  simulation,
  management_df = erahumed::albufera_management,
  ideal_flow_rate_cm = 5,
  height_thresh_cm = 2,
  seed = 840
)

Arguments

simulation

An object of class erahumed_simulation.

management_df

A data.frame that provides the yearly schedule for irrigation and draining, that strictly follows the template of albufera_management.

ideal_flow_rate_cm

A positive number. Ideal inflow/outflow of a cluster, for days in which the cluster is scheduled to be in flux (i.e. when being simultaneously irrigated and drained) Expressed in centimeters per day.

height_thresh_cm

A positive number. Height threshold for water levels, below which a cluster is considered to be emptied.

seed

A number. Seed for random number generation used by the algorithm.

Value

An object of class erahumed_simulation.

Details

The output data.frame for this model layer, retrieved through get_layer_output(layer = "hbp") has one row per cluster and per day, providing the simulated hydrological times-series for all paddy clusters. The data-set has the following columns:

date

Date of measurement

cluster_id

Cluster identifier.

area_m2

Surface area (in squared meters) of cluster.

tancat

TRUE or FALSE, whether the cluster is a tancat (cf. clusters).

variety

Variety of rice planted in the cluster. (cf. clusters).

area_m2

Surface area (in squared meters) of cluster.

ditch

Ditch to which the cluster pertains (cf. clusters).

capacity_m3_s

Outflow (to the Albufera Lake) of ditch.

seed_day

Whether date corresponds to the sowing day of the year.

petp_cm

Precipitation minus evapotranspiration per unit area, in centimeters (common to all clusters).

ideal_draining

TRUE or FALSE, whether the cluster is supposed to be drained on this day of the year.

ideal_irrigation

TRUE or FALSE, whether the cluster is supposed to be irrigated on this day of the year.

ideal_inflow_cm

Ideal inflow computed by the HBP algorithm (see TODO #46).

ideal_outflow_cm

Ideal outflow computed by the HBP algorithm (see TODO #46).

ideal_diff_flow_cm

Ideal inflow minus ideal outflow.

ideal_height_eod_cm

Ideal water level of cluster at the end of day (i.e. after draining and/or irrigation, for this day of the year.

draining

TRUE or FALSE, whether the cluster is actually (i.e. in the simulation) being drained on this day of the year.

irrigation

TRUE or FALSE, whether the cluster is actually ( i.e. in the simulation) being irrigated on this day of the year.

height_eod_cm

Simulated water level of cluster at the end of day.

height_sod_cm

Simulated water level of cluster at the end of day ( i.e. before irrigation and/or draining).

inflow_cm

Simulated inflow of cluster (in centimeters).

inflow_m3_s

Simulated inflow of cluster (in cube meters per second).

outflow_cm

Simulated outflow of cluster (in centimeters).

outflow_m3_s

Simulated outflow of cluster (in cube meters per second).

plan_delay

Accumulated delay on the yearly plan for this cluster. See (TODO #46) for a detailed description.

See also