INP: Input data

This model component has the only purpose to collect the observational (or, potentially, synthetic) data used by the algorithms of the ERAHUMED decision support system.

This is the first layer of the modeling chain and does not have any upstream dependence.

Usage

inp(model)

compute_inp(
  model,
  outflows_df = erahumed::albufera_outflows,
  weather_df = erahumed::albufera_weather
)

Arguments

model

An object of class erahumed_model.

outflows_df

A data.frame, whose structure follows the template of albufera_outflows; See details.

weather_df

A data.frame, whose structure follows the template of albufera_weather; See details.

Value

Objects of class erahumed_model and erahumed_inp, for compute_inp() and inp() respectively.

Details

The outflows_df input data.frame is meant to capture the observational hydrological data on the Albufera lake. This consists of a level column, that is the daily measured lake water level in meters, plus any number of columns named as outflow_* (e.g. outflow_pujol), that give the daily measured outflows in cube meters per second. The weather_df captures the relevant weather data for modeling:

Precipitation and evapotranspiration, relevant for modeling the hydrological balance, corresponding to the precipitation_mm and evapotranspiration_mm columns, respectively.

Temperature, relevant because it affects chemical reaction speeds.

Both data.frames should have a date column (of class Date), and the corresponding date domain should be an interval (i.e. no missing data between the maximum and minimum of date is allowed). The subsequent modeling will only be performed on the largest date interval in which both hydrological and weather input data is available.

See also

Author

Valerio Gherardi

Examples

model <- erahumed_model() |> compute_inp()
inp(model)
#> An ERAHUMED INP model component.
#> 
#> Output columns: date, level, outflow_pujol, outflow_perellonet, outflow_perello, is_imputed_level, is_imputed_outflow, temperature_ave, temperature_min, temperature_max, precipitation_mm, evapotranspiration_mm