Plotting local hydrological balance data

Plot method for local Hydrological balance data, generated through hb_local. A simple wrapper around plot_ly to generate time series plots of the calculated quantities.

Usage

# S3 method for class 'hb_local'
plot(x, type = c("cluster_levels", "map"), ...)

Arguments

x

The object of class hb_global containing the data to be plotted.

type

Type of plot to be generated. Currently, only the "cluster_levels" plot is implemented.

...

Further plotting parameters, associated with the plot type specified by type; See details.

Value

A plotly plot.

Details

The "cluster_levels" plot generates a time series plot of the daily water levels (in cm) of an individual cluster. In order to use this plotting method, the user must provide an additional cluster_id argument, a string specifying the identifier of the cluster whose levels are to be plotted (see example below).

Examples

plot(albufera_hb_local(),
     type = "cluster_levels",
     cluster_id = "02_Carrera_del_Saler0-2_0"
     )