HBA: Hydrological Balance of the Albufera Lake

This simulation layer completes the partial hydrological balance information provided as input data, by computing the total inflow to the Albufera lake, as well as, if relevant, the amount of unaccounted outflow - usually attributed to tancats that suck water from the lake.

Usage

setup_hba(
  simulation,
  storage_curve = function(level) 16.7459 * 1e+06 + level * 23.6577 * 1e+06,
  petp_function = function(p, etp) 114.226 * 1000 * p - 79.361 * 1000 * etp
)

Arguments

simulation

An object of class erahumed_simulation.

storage_curve

a function that takes a numeric vector as input, and returns a numeric vector of the same length. Function that converts lake levels into lake volumes.

petp_function

a function that takes two numeric vectors of common length as inputs, and returns a numeric vector of the same length. Function that converts precipitation and evapotranspiration values (assumed to be expressed in millimiters) into an overall volume change (in cube meters when the inputs are given in the expected units).

Value

An objects of class erahumed_simulation.

Details

The numeric inputs for the linear storage curve are taken from the CHJ report modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.. The values used as the arguments of petp_function() were calculated by the package authors, and correspond to the total study area (surface_P) and the flooded surface (surface_ETP).

The output data.frame of the calculated HBA layer contains a copy of the input data, as well as the following additional columns:

volume Volume time series, obtained from the storage curve.

volume_change Differenced time series of volume. The \(n\)-th is given by \(\Delta V _n \equiv V_{n+1}-V_n\), where \(V_n\) is volume at time step \(n\).

petp_change Time series. The amount \(\Delta V _n ^\text{P-ETP}\) of volume change due to precipitation and evapotranspiration, obtained from the P-ETP surface.

outflow_total Time series of total outflows, measured in cube meters per second. This is the sum \(\sum _i O_i\) of time series passed through the outflows argument, plus an extra term \(\delta O\) of unaccounted outflow. The correction term is defined by \(\delta O = \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60} -\sum _i O_i\) and is chosen in such a way to ensure that the total inflow is always non-negative.

outflow_extra Time series. The unaccounted outflow term \(\delta O\) described above.

inflow_total Time serie of total inflows, in cube meters per second. This is computed as \(I = \sum _{i} O_i + \delta O + \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60}\).

residence_time_days. Residence time, as simulationed by hba_residence_time.

See also

Author

Pablo Amador Crespo, Valerio Gherardi