Hydrological Balance of the Albufera lake

This layer computes the hydrological balance of the Albufera lake, complementing the input information on lake outflows and levels with an estimate of the total inflow from irrigation ditches.

Usage

setup_hba(
  simulation,
  storage_curve = function(level) 16.7459 * 1e+06 + level * 23.6577 * 1e+06,
  petp_function = function(p, etp) 114.226 * 1000 * p - 79.361 * 1000 * etp
)

Arguments

simulation

[erahumed_simulation]
The simulation object being modified.

storage_curve

[function]
A function that converts lake levels into lake volumes. The function should take a single numeric vector as input, and should return a numeric vector of the same length as output. The numeric values for the default (linear) storage curve are taken from the CHJ report Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.

petp_function

[function]
A function that converts values of precipitation and evapotranspiration per unit area (assumed to be expressed in millimiters) into volume changes for the Albufera lake (in cube meters when the inputs are given in the expected units). This function should take two numeric vectors of common length as inputs (representing the precipitation and evapotranspiration time series, respectively), and should return a numeric vector of the same length. The numeric values used in the default (linear) function were computed by the package authors, and correspond to the total study area (consisting of the lake and irrigation ditches) and the flooded fraction, that are relevant for precipitation and evapotranspiration volume changes, respectively.

Value

An objects of class erahumed_simulation.

Details

The output data.frame of the calculated HBA layer contains a copy of the input data, as well as the following additional columns:

volume Volume time series, obtained from the storage curve.

volume_change Differenced time series of volume. The \(n\)-th is given by \(\Delta V _n \equiv V_{n+1}-V_n\), where \(V_n\) is volume at time step \(n\).

petp_change Time series. The amount \(\Delta V _n ^\text{P-ETP}\) of volume change due to precipitation and evapotranspiration, obtained from the P-ETP surface.

outflow_total Time series of total outflows, measured in cube meters per second. This is the sum \(\sum _i O_i\) of time series passed through the outflows argument, plus an extra term \(O_R\) that estimates the outflow due to water recirculation (see TODO #60).

outflow_recirculation Time series. Estimate of outflow due to water recirculation.

inflow_total Time serie of total inflows, in cube meters per second. This is computed as \(I = \sum _{i} O_i + \delta O + \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60}\).

residence_time_days. Residence time, as simulationed by hba_residence_time.

See also