HBP: Hydrological Balance of rice Paddies

This model component performs a simulation of hydrological balance daily data for the rice paddies of the Albufera National Park, that provides a synthetic dataset of daily inflows, outflows and average water levels for each rice paddy.

The inputs to this simulation are:

The previously calculated hydrological balance for the Albufera lake.

A geographical subdivision of the Albufera National Park into "clusters" of rice paddies, of a definite rice variety, and whose waters flow into the same ditch.

An yearly ideal schedule for irrigation and draining, with corresponding expected water levels of a cluster for each day of the year.

This modeling layer requires the hba component of the model to be pre-computed.

Usage

hbp(model)

compute_hbp(
  model,
  management_df = erahumed::albufera_management,
  clusters_df = erahumed::albufera_clusters,
  ideal_flow_rate_cm = 5,
  height_thresh_cm = 2
)

Arguments

model

An object of class erahumed_model, with a pre-computed hba component (i.e. such that hba(model) is not NULL).

management_df

A data.frame that provides the yearly schedule for irrigation and draining, that strictly follows the template of albufera_management.

clusters_df

A data.frame that contains the definition of rice paddy clusters. This should strictly follow the template of albufera_clusters, and it is in fact unlikely that the user would want to change this input (TODO #47).

ideal_flow_rate_cm

A positive number. Ideal inflow/outflow of a cluster, for days in which the cluster is scheduled to be in flux (i.e. when being simultaneously irrigated and drained) Expressed in centimeters per day.

height_thresh_cm

A positive number. Height threshold for water levels, below which a cluster is considered to be emptied.

Value

Objects of class erahumed_model and erahumed_hbp, for compute_hbp() and hbp() respectively.

Details

TODO: #64, plus more detailed information on the structure of management_df (and perhaps also on clusters_df, depending on the decision taken regarding #47).

See also

Author

Pablo Amador Crespo, Valerio Gherardi